move to FOV-based normalization scheme
